% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{alert}
\alias{alert}
\title{User alert message}
\usage{
alert(...)
}
\arguments{
\item{...}{the alert message as string in the function}
}
\value{
prints the alert message.
}
\description{
User alert message
}
\examples{
alert("Encountered a problem")

}
