% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{areColors}
\alias{areColors}
\title{Find if Color Exists}
\usage{
areColors(x)
}
\arguments{
\item{x}{the name of the color}
}
\value{
returns TRUE if the color exist, FALSE if it does not.
}
\description{
Find if Color Exists
}
\examples{
areColors("blue")

}
