% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-SBP.R
\name{arg.as.char}
\alias{arg.as.char}
\title{Represent input argument as a character}
\usage{
arg.as.char(x)
}
\arguments{
\item{x}{the input argument}
}
\value{
returns the input argument as character string
}
\description{
Represent input argument as a character
}
\examples{
arg.as.char(binarydata)

}
