% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{bar.plot}
\alias{bar.plot}
\title{Create a SBP bar plot}
\usage{
bar.plot(y.clm, data, all = F, y.name = NULL, clr = NULL)
}
\arguments{
\item{y.clm}{the "y" column of the data set.}

\item{data}{the data set from which the data is taken.}

\item{all}{#ask_Stan}

\item{y.name}{the name of the "y" column.Default value is set to NULL.}

\item{clr}{The color(s)  in the plot(s). Default value is set to NULL.}
}
\value{
shows a bar plot of the "y" column data.
}
\description{
Create a SBP bar plot
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
bar.plot("len", data_frame)

}
