% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{box.plot}
\alias{box.plot}
\title{Create a SBP box plot}
\usage{
box.plot(input, data, y.name = NULL, clr = "rainbow")
}
\arguments{
\item{input}{the input as a column name in quotation marks or a formula.}

\item{data}{the data set which contains the input column(s).}

\item{y.name}{the name of the "y" column of the data set. Default value is set to NULL.}

\item{clr}{The color of the graph(s) in the plot(s).Default value is set to "rainbow".}
}
\value{
shows a box plot based on the input type(single column data or formula).
}
\description{
Create a SBP box plot
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
box.plot("len", data_frame)
box.plot(len~supp,  data_frame)

}
