% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.centers}
\alias{compare.centers}
\title{Compare centers of a numeric variable across groups}
\usage{
compare.centers(
  form,
  data,
  txt = 1,
  tbl = 1,
  fig = 1,
  name.y = NULL,
  name.grp = NULL,
  clr = "rainbow"
)
}
\arguments{
\item{form}{The formula which is used to define the relationship between the two variables.}

\item{data}{The data set from which the two variables are taken.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{name.y}{The 'y' variable name in the formula. Default value is NULL. #ask_Stan}

\item{name.grp}{The 'group' variable name in the formula. Default value is NULL. #ask_Stan}

\item{clr}{The color(s) used in the plot(s).}
}
\value{
describes the relationship between the variables with a  narrative,plots and references .
}
\description{
Compare centers of a numeric variable across groups
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
compare(len~supp,data_frame)

}
