% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.events}
\alias{compare.events}
\title{Compare events between 2 variables if the 'y' variable is of survival class.}
\usage{
compare.events(
  form,
  data,
  txt = 1,
  tbl = 1,
  fig = 1,
  name.y = NULL,
  name.grp = NULL,
  clr = "rainbow"
)
}
\arguments{
\item{form}{The formula which is used to define the relationship of the two variables.}

\item{data}{The data set from which the two variables are taken.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{name.y}{The 'y' variable name in the formula. Default value is NULL. #ask_Stan}

\item{name.grp}{The 'group' variable name in the formula. Default value is NULL. #ask_Stan}

\item{clr}{The color(s) used in the plot(s).}
}
\value{
describes the relationship between the variables with a narrative,plots and references.
}
\description{
Compare events between 2 variables if the 'y' variable is of survival class.
}
\examples{
#ask_Stan


}
