% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.proportions}
\alias{compare.proportions}
\title{Compare Proportions}
\usage{
compare.proportions(
  form,
  data,
  tbl = 1,
  fig = 1,
  txt = 1,
  y.name = NULL,
  grp.name = NULL,
  clr = "rainbow"
)
}
\arguments{
\item{form}{The formula which is used to define the relationship of the two variables.}

\item{data}{The dataset from which the two variables are taken.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{y.name}{The 'y' variable in the formula. Default value is NULL.}

\item{grp.name}{The 'group' variable in the formula. Default value is NULL.}

\item{clr}{The color of the graphs in the plot(s).}
}
\value{
The text,plots and references describing the relationship between the variables.
}
\description{
Compare Proportions
}
