% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{competing.event.time}
\alias{competing.event.time}
\title{Creating a competing event time variable
#probably needs a dataset as parameter.}
\usage{
competing.event.time(obs.time, obs.event, ev.key = NULL)
}
\arguments{
\item{obs.time}{the time variable in the data set.}

\item{obs.event}{the event variable in the data set.}

\item{ev.key}{the event key. The default value is set to NULL.}
}
\value{
returns a new variable combining observation time and observation event
}
\description{
Creating a competing event time variable
#probably needs a dataset as parameter.
}
\examples{
#ask_Stan
data_frame <- data.frame(time = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         event = c(0, 0, 1, 0, 1, 0, 1, 0, 0, 0),
                         Age = c(50, 55, 45, 49, 52, 43, 62, 60, 70, 71))
competing.event.time("time","event")

}
