% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{coxph.tbl}
\alias{coxph.tbl}
\title{Create a table of hazard ratio estimates for a coxph model #ask_Stan}
\usage{
coxph.tbl(coxph.result)
}
\arguments{
\item{coxph.result}{the result of the coxph model}
}
\value{
returns a data frame of hazard ratio estimates of the coxph model.
}
\description{
Create a table of hazard ratio estimates for a coxph model #ask_Stan
}
\examples{
#ask_Stan

}
