% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{crr.tbl}
\alias{crr.tbl}
\title{Create a table of coefficient estimates for a competing risk regression model}
\usage{
crr.tbl(crr.result)
}
\arguments{
\item{crr.result}{the coefficient result of the risk regression model.}
}
\value{
returns the coefficient estimates in a data frame format.
}
\description{
Create a table of coefficient estimates for a competing risk regression model
}
\examples{
#ask_Stan

}
