% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{cut.quantile}
\alias{cut.quantile}
\title{Categorize a numeric variable by quantile}
\usage{
\method{cut}{quantile}(x, n = 4)
}
\arguments{
\item{x}{name of the numeric variable}

\item{n}{the number of subgroup to create.By default, the division value is 4, meaning it will create quartile division.}
}
\value{
returns the variable data as quantile/subgroups.
}
\description{
Categorize a numeric variable by quantile
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
cut.quantile(data_frame$len)

}
