% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{cx.test}
\alias{cx.test}
\title{Test of independence on a cross-tabulation}
\usage{
cx.test(cx.tbl, B = 9999)
}
\arguments{
\item{cx.tbl}{#ask_Stan}

\item{B}{#ask_Stan}
}
\value{
#ask_Stan
}
\description{
\code{cx.test()} tests the independence on a cross-tabulation result.
}
\examples{
#ask_Stan

}
