% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{define.colors}
\alias{define.colors}
\title{Choose Colors}
\usage{
define.colors(n, clr.palette = "rainbow")
}
\arguments{
\item{n}{the number of colors to add}

\item{clr.palette}{source color name. Default value is set to "rainbow"}
}
\value{
displays n number of colors from the color palette
}
\description{
Choose Colors
}
\examples{
define.colors(5, clr.palette="rainbow")

}
