% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.categorical}
\alias{describe.categorical}
\title{Describe a categorical variable}
\usage{
describe.categorical(
  x,
  tbl = 1,
  fig = 1,
  txt = 1,
  clr = NULL,
  x.name = NULL,
  use.all = T
)
}
\arguments{
\item{x}{the categorical variable to describe.}

\item{tbl}{display the result in tabular format. The values are 0=none, 1=basic, 2=detailed.}

\item{fig}{display the result in figures. The values are 0=none, 1=basic, 2 and higher=more.}

\item{txt}{display narrative output. The values are 0=none, 1=basic, 2=detailed.}

\item{clr}{colors to use in displaying the figures.}

\item{x.name}{name of x variable to use in narrative output}

\item{use.all}{indicates whether to include all data regardless of missingness. Default value is True.}
}
\value{
returns description of the categorical variable with tables, figures, and a narrative.
}
\description{
Describe a categorical variable
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
describe("supp",data_frame)

}
