% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.event.timing}
\alias{describe.event.timing}
\title{Describe Survival Variable}
\usage{
describe.event.timing(x, tbl = 1, fig = 1, txt = 1, clr = NULL, x.name = NULL)
}
\arguments{
\item{x}{the time event data that is used for analysis.}

\item{tbl}{display the result in tabular format. The values are 0=none, 1=basic, 2=detailed.}

\item{fig}{display the result in figures. The values are 0=none, 1=basic, 2 and higher=more.}

\item{txt}{display narrative output. The values are 0=none, 1=basic, 2=detailed.}

\item{clr}{colors to use in displaying the figures.}

\item{x.name}{}
}
\value{
returns event time distribution with tables, figures, and a narrative.
}
\description{
Describe Survival Variable
}
\examples{
#ask_Stan

}
