% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe.numeric}
\alias{describe.numeric}
\title{Describe a numeric variable}
\usage{
describe.numeric(x, tbl = 1, fig = 2, txt = 1, clr = NULL, x.name = NULL)
}
\arguments{
\item{x}{the name of the variable in quotation marks.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 2.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{clr}{colors to use in displaying the plot(s).}

\item{x.name}{character string giving name of x variable}
}
\value{
returns a description of the numeric variable with tables, figures, and a narrative.
}
\description{
Describe a numeric variable
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
describe("len",data_frame)

}
