% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a Variable/Feature}
\usage{
describe(
  clm.name,
  data,
  tbl = 1,
  fig = 1,
  txt = 1,
  clr = NULL,
  y.name = NULL,
  use.all = T
)
}
\arguments{
\item{clm.name}{the name of the column in quotation marks.}

\item{data}{the name of the data set.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{clr}{The color in the plot(s). Default value is NULL.}

\item{y.name}{name of x in quotation marks for narrative. #ask_Stan}

\item{use.all}{indicates whether to include all data regardless of missingness. Default value is True. #ask_Stan}
}
\value{
returns statistical description of the variable in terms of tables, figures and narratives.
}
\description{
Describe a Variable/Feature
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
describe("len",data_frame)

}
