% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.events}
\alias{estimate.events}
\title{This function estimates the timing of events in a categorical variable in a
data frame.}
\usage{
estimate.events(x, tbl = 1, fig = 1, txt = 1, clr = NULL, x.name = NULL)
}
\arguments{
\item{x}{The categorical variable to describe.}

\item{tbl}{A flag indicating whether to display a table. Default is 1.}

\item{fig}{A flag indicating whether to display a figure. Default is 1.}

\item{txt}{A flag indicating whether to display text. Default is 1.}

\item{clr}{The color(s) to use.}

\item{x.name}{The name of the x variable to use in narrative output. Default
is NULL.}
}
\value{
The estimated timing of events.
}
\description{
This function estimates the timing of events in a categorical variable in a
data frame.
}
