% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate.pop.value}
\alias{estimate.pop.value}
\title{Estimate population values}
\usage{
estimate.pop.value(
  x,
  null = NULL,
  tbl = 1,
  fig = 1,
  txt = 1,
  clr = NULL,
  x.name = NULL
)
}
\arguments{
\item{x}{The categorical variable to describe.}

\item{null}{The hypothesized null value. Default is NULL.}

\item{tbl}{A flag indicating whether to display a table. Default is 1.}

\item{fig}{A flag indicating whether to display a figure. Default is 1.}

\item{txt}{A flag indicating whether to display text. Default is 1.}

\item{clr}{The color(s) to use.}

\item{x.name}{The name of the x variable to use in narrative output. Default
is NULL.}
}
\value{
The estimated population value.
}
\description{
This function estimates the population value of a categorical variable in a
data frame.
}
\examples{
data_frame <- data.frame(time = c(1,2,3,4,5,6,7,8,9,10),
                         y = c(11,12,13,14,15,16,17,18,19,20),
                         status = c("A","B","A","A","B","A","A","B","A","A"))
estimate.pop.value(x = data_frame$status)

}
