% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{event.plot}
\alias{event.plot}
\title{Create a SBP event plot #ask_Stan}
\usage{
event.plot(input, data, y.name = NULL, clr = NULL)
}
\arguments{
\item{input}{the input can be both a column with quatation mark or a formula.}

\item{data}{the name of the data set that contains the input column(s).}

\item{y.name}{name of the "y" column.}

\item{clr}{the color of the plot.}
}
\value{
shows an event plot based on the input type(formula or column data).
}
\description{
Create a SBP event plot #ask_Stan
}
\examples{
#ask_Stan

}
