% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{file.extension}
\alias{file.extension}
\title{Extract a file extension type}
\usage{
file.extension(file.name)
}
\arguments{
\item{file.name}{the name of the file with extension}
}
\value{
returns the extension name of the file
}
\description{
Extract a file extension type
}
\examples{
file.extension("binary.csv")

}
