% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit.cox}
\alias{fit.cox}
\title{Test the fit of a Cox regression model #ask_Stan}
\usage{
fit.cox(form, data)
}
\arguments{
\item{form}{the formula that contains the variables in the form 'y~x'.}

\item{data}{the data set name that contains the variables.}
}
\value{
returns the result of the cox regression model in a data frame format
}
\description{
Test the fit of a Cox regression model #ask_Stan
}
\examples{
#ask_Stan

}
