% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-SBP.R
\name{fresh.start}
\alias{fresh.start}
\title{Fresh R Session}
\usage{
fresh.start()
}
\value{
clears the R environment and starts new session with the source files and packages.
}
\description{
Fresh R Session
}
\examples{
fresh.start()

}
