% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-SBP.R
\name{get.arg}
\alias{get.arg}
\title{Get Argument Value}
\usage{
get.arg(call.string, arg.string)
}
\arguments{
\item{call.string}{the string that is called and converted into string}

\item{arg.string}{the string to match with}
}
\value{
returns ??
}
\description{
Get Argument Value
}
\examples{
#ask_Stan

}
