% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-SBP.R
\name{get.data.set}
\alias{get.data.set}
\title{Get Data Set from the R global environment}
\usage{
get.data.set(data.set)
}
\arguments{
\item{data.set}{the name of the data set}
}
\value{
returns the data set as a data frame from the R global environment
}
\description{
Get Data Set from the R global environment
}
\examples{
#error_in_code

}
