% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{get.vars}
\alias{get.vars}
\title{Get the dependent(y) and independent(x) variables from a formula}
\usage{
get.vars(form)
}
\arguments{
\item{form}{The formula which defines the relationship of the variables.}
}
\value{
returns the x and y variables from the formula.
}
\description{
Get the dependent(y) and independent(x) variables from a formula
}
\examples{
get.vars(points~team)

}
