% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-SBP.R
\name{get.y.clm}
\alias{get.y.clm}
\title{Get Column Data}
\usage{
get.y.clm(y, dset)
}
\arguments{
\item{y}{the exact column name or close match to its name}

\item{dset}{the data set from which the column data will be extracted}
}
\value{
returns the 'y' column data from the data set
}
\description{
Get Column Data
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
get.y.clm(Supp,data_frame)

}
