% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{get.yvar}
\alias{get.yvar}
\title{Get y-variable from a formula}
\usage{
get.yvar(form)
}
\arguments{
\item{form}{The formula which defines the relationship of the variables.}
}
\value{
returns the y variable of the formula.
}
\description{
Get y-variable from a formula
}
\examples{
get.yvar(height~age)

}
