% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{glm.tbl}
\alias{glm.tbl}
\title{Create a table of coefficient estimates for a glm(generalized linear model)}
\usage{
glm.tbl(glm.result)
}
\arguments{
\item{glm.result}{the result of the glm.}
}
\value{
returns the glm result in a data frame format.
}
\description{
Create a table of coefficient estimates for a glm(generalized linear model)
}
\examples{
#ask_Stan

}
