% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.numeric}
\alias{model.numeric}
\title{Create a model for numeric data #ask_Stan}
\usage{
model.numeric(form, data, tbl = 1, fig = 3, txt = 1, clr = c("black", "red"))
}
\arguments{
\item{form}{the formula that contains the variables in the form 'y~x'.}

\item{data}{the name of the data set.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 3.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{clr}{the color(s) used in the plot(s).Default value is set to 'black' and 'red'.}
}
\value{
returns the model results with tables,plots and a narrative.
}
\description{
Create a model for numeric data #ask_Stan
}
\examples{
#ask_Stan

}
