% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Describe a model}
\usage{
model(form, data, tbl = 1, fig = 1, txt = 1, clr = NULL)
}
\arguments{
\item{form}{the formula that contains the variables in the form 'y~x'.}

\item{data}{the data set name that contains the variables.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{clr}{The color(s) to be used in the plot(s).}
}
\value{
returns the model results in tables,figures and a narrative.
}
\description{
Describe a model
}
\examples{
#ask_Stan
}
