% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.txt}
\alias{model.txt}
\title{Generate narrative results for a model}
\usage{
model.txt(model.result)
}
\arguments{
\item{model.result}{the result of the model ?? #ask_Stan}
}
\value{
return the model results in a text narrative depending on the model type.
}
\description{
\code{model.txt()} generates a text based narrative result of the model.
}
\examples{
##ask_Stan

}
