% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{monotone.rank.model}
\alias{monotone.rank.model}
\title{Linear Regression on Ranks Model #ask_Stan}
\usage{
monotone.rank.model(x, y)
}
\arguments{
\item{x}{#ask_Stan}

\item{y}{#ask_Stan}
}
\value{
#ask_Stan
}
\description{
Linear Regression on Ranks Model #ask_Stan
}
\examples{
#ask_Stan

}
