% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{mosaic.plot}
\alias{mosaic.plot}
\title{Create a SBP mosaic plot}
\usage{
mosaic.plot(form, data, clr = "rainbow", y.name = NULL, grp.name = NULL)
}
\arguments{
\item{form}{The formula which is used to compare categorical data with a numeric variable.}

\item{data}{name of the data set where the variables are taken.}

\item{clr}{the color of the graphs in the plot. Default value is set to "rainbow".}

\item{y.name}{The name of the 'y' variable in the formula. Default value is set to NULL.}

\item{grp.name}{The group variable name in the formula. Default value is set to NULL.}
}
\value{
displays a mosaic graph in different ways(with missing data,missing data as a distinct category or without missing data).
}
\description{
Create a SBP mosaic plot
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
mosaic.plot(len~supp, data_frame)

}
