% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-SBP.R
\name{normality.test}
\alias{normality.test}
\title{Perform a Normality Test}
\usage{
normality.test(x)
}
\arguments{
\item{x}{the feature in the data set whose distribution will be tested.}
}
\value{
returns the p-value of the feature that implies the distribution of the data.
}
\description{
Perform a Normality Test
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
normality.test(data_frame$len)

}
