% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Find Outliers in a numeric variable}
\usage{
outliers(clm, dset, y.name = NULL, fig = 1, txt = 1, clr = c("gray", "red"))
}
\arguments{
\item{clm}{the numeric variable to check for Outliers.}

\item{dset}{the data set name that has the column data.}

\item{y.name}{the name of the column variable.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{clr}{The color of the text. Default value is 'black'.}
}
\value{
shows if any outliers exists in the numeric variable with texts,figure and references.
}
\description{
Find Outliers in a numeric variable
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
outliers("len",data_frame)

}
