% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{pie.plot}
\alias{pie.plot}
\title{Create a SBP pie plot}
\usage{
pie.plot(y.clm, data, y.name = NULL, all = F, clr = NULL)
}
\arguments{
\item{y.clm}{the numerical variable in quotations from data set}

\item{data}{name of the data set}

\item{y.name}{the name of the 'y' variable in the formula. Default value is NULL. #ask_Stan}

\item{all}{#ask_Stan}

\item{clr}{The color(s) of the graph in the plot.Default value is set to NULL.}
}
\value{
shows a pie plot in different ways(missing data,no missing data and missing data in a distinct category)
}
\description{
Create a SBP pie plot
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
pie.plot("len", data_frame)
}
