% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SBP.result}
\alias{print.SBP.result}
\title{Print a function's result in a narrative}
\usage{
\method{print}{SBP.result}(sbp.result, knt = T, tbl = T, txt = T, method = T, ref = T)
}
\arguments{
\item{sbp.result}{The variable that stores the calculation results.}

\item{knt}{A flag that identifies if knitr is available. Default value is set to TRUE. #ask_Stan}

\item{tbl}{A flag to show tables in the result. Default value is set to TRUE.}

\item{txt}{A flag to show texts in the result. Default value is set to TRUE.}

\item{method}{#ask_Stan}

\item{ref}{A flag to show references in the result. Default value is set to TRUE.}
}
\value{
shows a functions results containing tables, results, methods and references.
}
\description{
Print a function's result in a narrative
}
\examples{
#ask_Stan

}
