% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read.data}
\alias{read.data}
\title{Read data file of different extensions}
\usage{
read.data(file.name = NULL)
}
\arguments{
\item{file.name}{the name of the file with extension.}
}
\value{
returns the data of the file as a data frame.
}
\description{
Read data file of different extensions
}
\examples{
read.data("binary.csv")

}
