% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{SBP.example.Rmd.slides}
\alias{SBP.example.Rmd.slides}
\title{Print SBP results in power point slides}
\usage{
SBP.example.Rmd.slides(SBP.code.string, slide.title)
}
\arguments{
\item{SBP.code.string}{#ask_Stan}

\item{slide.title}{#ask_Stan}
}
\value{
writes the SBP results in power point slides
}
\description{
Print SBP results in power point slides
}
\examples{
#ask_Stan


}
