% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{scatter.plot}
\alias{scatter.plot}
\title{Create a SBP scatter plot}
\usage{
scatter.plot(
  form,
  data,
  clr = c("black", "red"),
  x.name = NULL,
  y.name = NULL,
  line = NA,
  txt = 0
)
}
\arguments{
\item{form}{the formula the takes two numerical variables.}

\item{data}{The data frame that contains the 2 variables.}

\item{clr}{the color(s) of the plot. Default value is set to "black" and "red".}

\item{x.name}{the name of the "x" column. Default value is set to NULL.}

\item{y.name}{the name of the "y" column. Default value is set to NULL.}

\item{line}{the line to draw in the plot. Default value is set to NA.}

\item{txt}{A flag that indicates to display text. Default value is set to 0.}
}
\value{
shows a scatter plot displaying the relationship between two numeric variables.
}
\description{
Create a SBP scatter plot
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
scatter.plot(len~dose, data_frame)

}
