% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show.palettes}
\alias{show.palettes}
\title{Display Color Palette}
\usage{
show.palettes(n.colors = 8)
}
\arguments{
\item{n.colors}{the number to color to show in each palette. The default value of n.colors is set to 8.}
}
\value{
displays the n.colors in each palette
}
\description{
Display Color Palette
}
\examples{
show.palettes()

}
