% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{stop.alert}
\alias{stop.alert}
\title{Alert user and stop calculation}
\usage{
stop.alert(...)
}
\arguments{
\item{...}{the error message as string in the function}
}
\value{
print the alert message and stops execution of program.
}
\description{
Alert user and stop calculation
}
\examples{
stop.alert("found error: program will halt")

}
