% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{text.list}
\alias{text.list}
\title{Write list items as a narrative}
\usage{
\method{text}{list}(x)
}
\arguments{
\item{x}{the vector containing the list items.}
}
\value{
returns the list items as a narrative.
}
\description{
Write list items as a narrative
}
\examples{
x=c("result 1","result 2","result 3")
text.list(x)

}
