% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{word.table}
\alias{word.table}
\title{Paste the table results in a word file}
\usage{
word.table(SBP.result)
}
\arguments{
\item{SBP.result}{the format of the result.}
}
\value{
outputs the result tables in a word file.
}
\description{
Paste the table results in a word file
}
\examples{
#ask_Stan

}
