% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{Compare two variables based on their type}
\usage{
associate(
  forms,
  data,
  txt = 1,
  tbl = 1,
  fig = 1,
  y.name = NULL,
  x.name = NULL,
  clr = NULL,
  line = 1
)
}
\arguments{
\item{line}{}
}
\description{
Compare two variables based on their type
}
