% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare.proportions}
\alias{compare.proportions}
\title{Compare Proportions between variables when the 'y' variable is of factor class.}
\usage{
compare.proportions(
  form,
  data,
  tbl = 1,
  fig = 1,
  txt = 1,
  y.name = NULL,
  grp.name = NULL,
  clr = "rainbow"
)
}
\arguments{
\item{form}{The formula which is used to define the relationship between the two variables.}

\item{data}{The data set from which the two variables are taken.}

\item{tbl}{A flag that indicates to display the number of table(s). Default value is 1.}

\item{fig}{A flag that indicates to display the number of figure(s). Default value is 1.}

\item{txt}{A flag that indicates to display text. Default value is 1.}

\item{y.name}{The 'y' variable name in the formula. Default value is NULL. #ask_Stan}

\item{grp.name}{The 'group' variable name in the formula. Default value is NULL. #ask_Stan}

\item{clr}{The color(s) used in the plot(s).}
}
\value{
describes the relationship between the variables with a narrative,plots and references.
}
\description{
Compare Proportions between variables when the 'y' variable is of factor class.
}
\examples{
data_frame <- data.frame(len = c(11.2, 8.2, 10.0, 27.3, 14.5, 26.4, 4.2, 15.2, 14.7, 10.4),
                         supp = c("VC","OJ","VC","VC","VC","OJ","VC","OJ","VC","OJ"),
                         dose = c(0.5, 0.5, 0.5, 2.0, 1.5, 1.0, 1.0, 2.0, 0.5, 2.0))
compare(supp~len,data_frame)

}
